disconnect settings changed signals in Sensors::destroy()
When disabling and re-enabling the extension, the old Sensors object is
destroyed and a new one is created. However, I forgot to disconnect the
settings "changed" signal handlers in the destroyed Sensors object.

So when changing some nvidia-smi related settings after having
disabled and re-enabled the extension, the old settings listeners would
still be active and react to the changed settings by re-spawning an
nvidia-smi process. This could lead to numerous nvidia-smi processes
being active at the same time, and related resource leaks.